+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Electrum parser: added symbol to worksheet name instead of wallet name.
 - Blockscout parser: set the worksheet name the same as the wallet name.
 - Price tool: added rate-limiter, retries and backoff handling for each API to prevent failures.
+- Conversion tool: OutputExcel determines if macOS via parameter instead of direct from platform.
 
 ## Version [0.6.0] (2025-11-05)
 Important:-